changed all labels to int

+++ Entropy_Splitter.cpp
@@ -2,7 +2,7 @@
  * @ Author: Adam Myers
  * @ Create Time: 2023-04-11 11:36:17
  * @ Modified by: Adam Myers
- * @ Modified time: 2023-04-11 15:17:29
+ * @ Modified time: 2023-04-13 17:34:15
  * @ Description: Header file for the Entropy_Splitter derived class.
  * 
  * This class inherets all the utility functions of the Splitter base class 
@@ -14,7 +14,7 @@
 #include "Entropy_Splitter.h"
 
 Entropy_Splitter::~Entropy_Splitter() {};
-double Entropy_Splitter::calculate_node_score(const std::vector<std::pair<std::vector<float>, float>>& data, const std::vector<size_t> indexes) const
+double Entropy_Splitter::calculate_node_score(const std::vector<std::pair<std::vector<float>, int>>& data, const std::vector<size_t> indexes) const
     {
             std::unordered_map<int, std::size_t> label_counts_map = label_counts(data, indexes);
             double n_data_points = indexes.size();
