+++ Entropy_splitter.h
@@ -2,7 +2,7 @@
  * @ Author: Adam Myers
  * @ Create Time: 2023-04-11 11:36:25
  * @ Modified by: Adam Myers
- * @ Modified time: 2023-04-11 15:19:38
+ * @ Modified time: 2023-04-13 17:34:19
  * @ Description: Implementation of the Entropy_Splitter derived class.
  */
 
@@ -15,7 +15,7 @@ class Entropy_Splitter : public Splitter
 {
 public:
     virtual ~Entropy_Splitter();
-    double calculate_node_score(const std::vector<std::pair<std::vector<float>, float>>& data, const std::vector<size_t> indexes) const;
+    double calculate_node_score(const std::vector<std::pair<std::vector<float>, int>>& data, const std::vector<size_t> indexes) const;
 };
 
 #endif
